Bug fixes
When user didn't supply a path as a cmd line arg and intended to use the current directory, the script incorrectly set up the git repository in the user's home directory. Also, when a user did specify a path, script would only use the path if it was in a directory that didn't exist yet (i.e., it made the current directory the git repo unintentionally).